Notmuch: Fixed modify-labels command on folders
When calling `<modify-labels>` on non-virtual folders with the toggle
operator ("!") the notmuch code would break early and leave the rest of
the arguments within the input buffer. Mutt would then interpret the "!"
as OP_SHELL_ESCAPE and pass the label to /bin/sh.

To fix this behaviour two changes have been made:

- When not within a virtual folder the error will now be displayed using
  `mutt_error` which causes mutt to stop executing the current macro.
  Without this change other instructions in the macro would be called.
  e.g. causing the ui to change, reload of inboxes etc.

- Remaining arguments in the for the <modify-labels> command will be
  discarded.